Decompose a helper for cloning process.env for npm-related commands.
